refactor: extract context pack utilities
